+++ R/digitizeRadii.R
@@ -4,7 +4,7 @@
 #' 
 #' The user can interactively select points on a plot by clicking with the first mouse button on the image. When finished, the x- and y-coordinates for each selected point will be returned.
 #' 
-#' @param fname A string that indicates the image (must be PNG, JPG, or BMP) to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).
+#' @param img A string that indicates the image (must be PNG, JPG, or BMP) to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).
 #' @param id A unique identifier for the fish or structure being examined. Will be coerced to a character.
 #' @param reading See details in \code{\link{RFBCoptions}}.
 #' @param description See details in \code{\link{RFBCoptions}}.
@@ -12,6 +12,7 @@
 #' @param edgeIsAnnulus See details in \code{\link{RFBCoptions}}.
 #' @param sepWindow See details in \code{\link{RFBCoptions}}.
 #' @param windowSize See details in \code{\link{RFBCoptions}}.
+#' @param popID See details in \code{\link{RFBCoptions}}.
 #' @param scaleBar See details in \code{\link{RFBCoptions}}.
 #' @param scaleBarLength See details in \code{\link{RFBCoptions}}.
 #' @param col.scaleBar See details in \code{\link{RFBCoptions}}.
@@ -24,6 +25,10 @@
 #' @param pch.sel See details in \code{\link{RFBCoptions}}.
 #' @param col.sel See details in \code{\link{RFBCoptions}}.
 #' @param cex.sel See details in \code{\link{RFBCoptions}}.
+#' @param showInfo See details in \code{\link{RFBCoptions}}.
+#' @param pos.info See details in \code{\link{RFBCoptions}}.
+#' @param cex.info See details in \code{\link{RFBCoptions}}.
+#' @param col.info See details in \code{\link{RFBCoptions}}.
 #'
 #' @details This uses \code{\link[graphics]{locator}} which will suspend activity in the console until the user indicates that they are done selecting points. The user indicates that they are done selecting points by right-clicking on the image and selecting Stop or selecting Stop in the window's menu.
 #'
@@ -46,13 +51,14 @@
 #' @examples
 #' ## None yet
 #' 
-digitizeRadii <- function(fname,id,reading,suffix,
-                          description,edgeIsAnnulus,
+digitizeRadii <- function(img,id,reading,suffix,
+                          description,edgeIsAnnulus,popID,
                           sepWindow,windowSize,scaleBar,
                           scaleBarLength,col.scaleBar,lwd.scaleBar,
                           scalingFactor,addTransect,snap2Transect,
                           col.transect,lwd.transect,
-                          pch.sel,col.sel,cex.sel) {
+                          pch.sel,col.sel,cex.sel,
+                          showInfo,pos.info,cex.info,col.info) {
   ## Process argument defaults
   if (missing(reading)) reading <- iGetopt("reading")
   if (missing(description)) description <- iGetopt("description")
@@ -60,6 +66,8 @@ digitizeRadii <- function(fname,id,reading,suffix,
   if (is.null(suffix) & !is.null(reading)) suffix <- reading
   if (missing(edgeIsAnnulus)) edgeIsAnnulus <- iGetopt("edgeIsAnnulus")
   if (!is.logical(edgeIsAnnulus)) STOP("'edgeIsAnnulus' must be TRUE or FALSE")
+  if (missing(popID)) popID <- iGetopt("popID")
+  if (!is.logical(popID)) STOP("'popID' must be TRUE or FALSE")
   if (missing(scaleBar)) scaleBar <- iGetopt("scaleBar")
   if (missing(scaleBarLength)) scaleBarLength <- iGetopt("scaleBarLength")
   if (scaleBar & is.null(scaleBarLength)) STOP("Must provide a 'scaleBarLength' when 'scaleBar=TRUE'.")
@@ -75,16 +83,21 @@ digitizeRadii <- function(fname,id,reading,suffix,
   if (missing(cex.sel)) cex.sel <- iGetopt("cex.sel")
   if (missing(sepWindow)) sepWindow <- iGetopt("sepWindow")
   if (missing(windowSize)) windowSize <- iGetopt("windowSize")
-  fname <- iHndlfname(fname)
-  id <- iHndlID(id)
+  if (missing(showInfo)) showInfo <- iGetopt("showInfo")
+  if (missing(pos.info)) pos.info <- iGetopt("pos.info")
+  if (missing(cex.info)) cex.info <- iGetopt("cex.info")
+  if (missing(col.info)) col.info <- iGetopt("col.info")
+  inames <- iHndlFilename(img)
+  id <- iHndlID(id,inames,popID)
 
-  ## Loads image given in fname
-  windowInfo <- iReadImage(fname$fname,sepWindow,windowSize)
-  message("** Loaded the ",fname$bn," image.")
+  ## Loads image given in img
+  windowInfo <- iReadImage(inames$givennm,id,sepWindow,windowSize,
+                           showInfo,pos.info,cex.info,col.info)
+  message("** Loaded the ",inames$basenm," image.")
   ## Allows the user to select a scaling bar to get a scaling factor
   SF <- iHndlScalingFactor(scaleBar,scaleBarLength,scalingFactor,
                            col.scaleBar,lwd.scaleBar,
-                        pixW2H=windowInfo$pixW2H)
+                           pixW2H=windowInfo$pixW2H)
   ## User selects annuli on the image
   pts <- iSelectAnnuli(pch.pts=pch.sel,col.pts=col.sel,
                        cex.pts=cex.sel,addTransect=addTransect,
@@ -92,13 +105,13 @@ digitizeRadii <- function(fname,id,reading,suffix,
   ## If asked, "snap" the selected points to the transect
   pts <- iSnap2Transect(pts)
   ## Converts the selected points to radial measurements
-  dat <- iProcessAnnuli(fname,pts,id,reading,suffix,description,
+  dat <- iProcessAnnuli(inames,pts,id,reading,suffix,description,
                         edgeIsAnnulus,SF$scalingFactor,
                         pixW2H=windowInfo$pixW2H)
   ## Add windowSize and scaling factor information to dat list
   dat <- c(dat,SF,windowInfo)
   ## Write the dat object to R object filename in the working directory.
-  save(dat,file=paste0(fname$dn,"/",dat$datobj))
+  save(dat,file=paste0(inames$dirnm,"/",dat$datobj))
   message("** All results written to ",dat$datobj)
   ## Invisibly return the R object
   invisible(dat)